Add experimental SIG Build RBE + toolchain configuration
I'm experimenting with a toolchain based specifically on the SIG Build
dockerfiles. Since they were created to match these pre-existing
toolchains, their configuration is actually the same for the most part,
except I've ripped out all of the conditional options to instead specify
env parameters explicitly.

The platform is called "sigbuild-r2.9". My line of thinking is:
internally, the DevInfra team only uses one RBE configuration at a time,
which is currently based on Python 3.9. tensorflow/toolchains must
retain previous *toolchain* configurations (not RBE configurations) for
future patch releases of TensorFlow. Therefore, much like the way the
containers are always tagged with the upcoming release version, I figure
we can continuously update the latest+1 toolchain with however we're
building at master HEAD, and once the branch is cut, create a new config
and use it in the dockerfiles instead.

For instance, the next version of TF is 2.9. The dockerfiles therefore
reference the "@sigbuild-r2.9" toolchain, which is configured here to
use CUDA 11.2. Whenever the containers are updated, they get pushed to
the "latest" and "2.9" tags. After the "next" version of TF changes to
2.10 when the branch is cut, we'd create a new "@sigbuild-r2.10"
toolchain here (with, say, CUDA 11.4, but it doesn't need to change) or
promote an ongoing testing toolchain, and change the toolchain
referenced in the dockerfiles. That way, the old 2.9 containers -- which
are now static -- correctly reference the 2.9 toolchain, which remains
available.

That's the idea, anyway; I need to do more testing to see if this will
work out correctly.